# src/generatedSpringRestApi/java/com/github/tuczv/addressBook/api/ChatV1Api.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.github.tuczv.addressBook.api;

import com.github.tuczv.addressBook.api.dto.ErrorApiDto;
import com.github.tuczv.addressBook.api.dto.MessageApiDto;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Tag(name = "ChatV1", description = "the ChatV1 API")
@RequestMapping("${openapi.addressbook.base-path:}")
public interface ChatV1Api {

    /**
     * POST /v1/messages : Post a message.
     *
     * @return OK (status code 200)
     *         or error (status code 400)
     */
    @Operation(
        operationId = "createMessage",
        summary = "Post a message.",
        tags = { "chat-v1" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MessageApiDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorApiDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/v1/messages",
        produces = { "application/json" }
    )
    ResponseEntity<MessageApiDto> createMessage(
        
    );


    /**
     * GET /v1/messages : Get a messages. Results are pageable with seek method.
     *
     * @return OK (status code 200)
     *         or error (status code 400)
     */
    @Operation(
        operationId = "findMessages",
        summary = "Get a messages. Results are pageable with seek method.",
        tags = { "chat-v1" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MessageApiDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorApiDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/v1/messages",
        produces = { "application/json" }
    )
    ResponseEntity<List<MessageApiDto>> findMessages(
        
    );

}