# src/openapi/addressbook-api.yaml
openapi: 3.0.0
info:
  title: Addressbook API
  description: |
    **Maintainer**: Cyd88
  version: 1.0.0
  contact:
    name: victor
    email: tuczkovski@gmail.com
servers:
  - url: /
paths:
  /v1/contacts:
    get:
      tags: [ contacts-v1 ]
      operationId: findContacts
      summary: Get contacts. Results are pageable with seek method.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'
        '400':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [ contacts-v1 ]
      summary: Add new contact.
      operationId: createContact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contact'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Contact'
        '400':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/contacts/{contactId}:
    get:
      tags: [ contacts-v1 ]
      operationId: findContactById
      summary: Get a contact by id. Results are pageable with seek method.
      parameters:
        - name: contactId
          in: path
          required: true
          description: Contact Id
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [ contacts-v1 ]
      summary: Update a contact by id.
      operationId: updateContactById
      parameters:
        - name: contactId
          in: path
          required: true
          description: Contact Id
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contact'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [ contacts-v1 ]
      operationId: removeContactById
      summary: Delete a contact by id.
      parameters:
        - name: contactId
          in: path
          required: true
          description: Contact Id
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/contacts/group/{groupName}:
    get:
      tags: [ contacts-v1 ]
      operationId: findContactByGroup
      summary: Get a contacts by group. Results are pageable with seek method.
      parameters:
        - name: groupName
          in: path
          required: true
          description: Group name
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'
        '400':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/messages:
    get:
      tags: [ chat-v1 ]
      operationId: findMessages
      summary: Get a messages. Results are pageable with seek method.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '400':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [ chat-v1 ]
      operationId: createMessage
      summary: Post a message.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/emails:
    get:
      tags: [ emails-v1 ]
      operationId: findEmails
      summary: Get a emails. Results are pageable with seek method.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Mail'
        '400':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [ emails-v1 ]
      operationId: createEmail
      summary: Post a email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Mail'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mail'
        '400':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/emails/{emailId}:
    delete:
      tags: [ events-v1 ]
      operationId: findEmailById
      summary: Get a email by Id.
      parameters:
        - name: emailId
          in: path
          required: true
          description: Email id
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mail'
        '400':
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      required: [ id, code, timestamp ]
      properties:
        id:
          type: string
          format: uuid
          description: Unique error ID, used for logging purposes, UUID format
        code:
          type: string
          description: A string coding the error type. This is given to caller so he can translate them if required.
        timestamp:
          type: integer
        message:
          type: string
          description: A short localized string that describes the error.
    Contact:
      title: Contact
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        group:
          type: string
        address:
          type: string
        user:
          type: string
...
    Mail:
      title: Mail
      type: object
      properties:
        id:
          type: string
          format: uuid
        userTo:
          type: string
        subject:
          type: string
        body:
          type: string
        date:
          type: string
        userFrom:
          type: object
    Message:
      title: Mail
      type: object
      properties:
        id:
          type: string
          format: uuid
        author:
          type: string
        text:
          type: string
        date:
          type: string
        subject:
          type: string